RAG/Siddhardhan/requirements.txt
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
#langchain-groq
langchain-ollama
unstructured
unstructured[pdf]
nltk
sentence-transformers
pypdf
python-dotenv
streamlit
uvicorn